rename: search-context.tsx → searchContext.tsx for camelCase consistency
